% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_eval.R
\name{answer_fn_with_env}
\alias{answer_fn_with_env}
\title{Create an answer checking function that retains its environment}
\usage{
answer_fn_with_env(fn, label = NULL)
}
\arguments{
\item{fn}{Function for checking the user's input; must provide one argument (user input).}

\item{label}{Optional answer label.}
}
\value{
Object of class `tutorial_question_answer` with the answer function's code as `value`.
}
\description{
Evaluate the users's submission to determine correctness and to return feedback using a function `fn` which
expects as only argument the user's input as character string. Any atomic objects from `fn`'s environment are
retained.
}
