% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_html_dependencies}
\alias{append_html_dependencies}
\title{Append necessary HTML dependencies (JS, CSS)}
\usage{
append_html_dependencies(html_dependencies, doc_config, introjs = FALSE)
}
\arguments{
\item{html_dependencies}{list of existing HTML dependencies}

\item{doc_config}{configuration options that will be embedded as JSON in the document}

\item{introjs}{if TRUE, include the intro.js library}
}
\value{
updated list of HTML dependencies
}
\description{
Appends all necessary HTML dependencies (JS libraries, CSS, etc.) to the list of existing HTML dependencies
given as `html_dependencies`.
}
\keyword{internal}
