% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial-format.R
\name{check_algebraic_expression}
\alias{check_algebraic_expression}
\title{Helper function to be used to check numeric algebraic expressions like "(1 + 3) / 3.14" in a [learnr::answer_fn()].}
\usage{
check_algebraic_expression(
  expected_result,
  allowed_chars = "0123456789\\\\.\\\\+\\\\-\\\\*/\\\\(\\\\)",
  allowed_max_length = 128
)
}
\arguments{
\item{expected_result}{the expected result after evaluating `input`}

\item{allowed_chars}{all characters that are allowed in `input`}

\item{allowed_max_length}{allowed maximum string length of `input`}
}
\value{
a function to be used in [learnr::answer_fn()]
}
\description{
Dismisses all inputs that are longer than `allowed_max_length` or contain characters other than those in
`allowed_chars`. The latter is very important as the input will be evaluated by the R interpreter and one should
watch out that it only contains numeric algebraic expressions.
}
