% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_eval.R
\name{fn_text_with_envvars_injected}
\alias{fn_text_with_envvars_injected}
\title{Helper function to inject variables of `fn`'s environment directly into the function body of `fn`
and return the function's code as text. Only atomic objects from `fn`'s environment are
considered.}
\usage{
fn_text_with_envvars_injected(fn)
}
\arguments{
\item{fn}{Function to turn into code text.}
}
\value{
Code of `fn` injected with the function's environment variables.
}
\description{
Helper function to inject variables of `fn`'s environment directly into the function body of `fn`
and return the function's code as text. Only atomic objects from `fn`'s environment are
considered.
}
