% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_includes}
\alias{get_includes}
\title{Get HTML files (HTML snippets) to include}
\usage{
get_includes(consentmodal = TRUE, dataprotectmodal = TRUE, feedback = TRUE)
}
\arguments{
\item{consentmodal}{either logical value or string pointing to an HTML file for a "data protection consent" modal}

\item{dataprotectmodal}{either logical value or string pointing to an HTML file for a "data protection license" modal}

\item{feedback}{either logical value or string pointing to an HTML file for a feedback element}
}
\value{
list with includes
}
\description{
Get HTML files (HTML snippets) to include
}
\keyword{internal}
