% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_eval.R
\name{question_mathexpression_percentage}
\alias{question_mathexpression_percentage}
\title{Quiz question that allows to provide an answer with a simple mathematical expression
that is being evaluated and checked against `expected_result`. Expect a percentage as answer.}
\usage{
question_mathexpression_percentage(
  text,
  expected_result,
  placeholder = "Enter a mathematical expression that evaluates to a percentage ...",
  ...
)
}
\arguments{
\item{text}{Question or option text}

\item{expected_result}{expected result (numeric)}

\item{placeholder}{A character string giving the user a hint as to what can
be entered into the control. Internet Explorer 8 and 9 do not support this
option.}

\item{...}{Answers created with \code{\link[learnr:answer]{answer()}} or \code{\link[learnr:answer_fn]{answer_fn()}}, or extra
parameters passed onto \code{\link[learnr:question]{question()}}. Answers with custom function checking}
}
\description{
Quiz question that allows to provide an answer with a simple mathematical expression
that is being evaluated and checked against `expected_result`. Expect a percentage as answer.
}
