% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_eval.R
\name{question_text_custom_answer_fn}
\alias{question_text_custom_answer_fn}
\title{Text-based quiz question that allows to check the answer using a custom answer function `answer_fn`.
Contrary to `learnr::question_text`, all atomic objects in `answer_fn`'s environment are retained.}
\usage{
question_text_custom_answer_fn(
  text,
  answer_fn,
  placeholder = "Provide your answer here ...",
  trim = TRUE,
  ...
)
}
\arguments{
\item{text}{Question or option text}

\item{answer_fn}{function for checking the user's input; must provide one argument (user input)}

\item{placeholder}{A character string giving the user a hint as to what can
be entered into the control. Internet Explorer 8 and 9 do not support this
option.}

\item{trim}{Logical to determine if whitespace before and after the answer
should be removed.  Defaults to \code{TRUE}.}

\item{...}{Answers created with \code{\link[learnr:answer]{answer()}} or \code{\link[learnr:answer_fn]{answer_fn()}}, or extra
parameters passed onto \code{\link[learnr:question]{question()}}. Answers with custom function checking}
}
\description{
Text-based quiz question that allows to check the answer using a custom answer function `answer_fn`.
Contrary to `learnr::question_text`, all atomic objects in `answer_fn`'s environment are retained.
}
