% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{survey}
\alias{survey}
\title{Generate a survey from survey item definitions passed as list `items`.}
\usage{
survey(items, caption = "Survey", message = "Thank you.")
}
\arguments{
\item{items}{Item definitions. Must be a list, where each list element in turn is a list that defines a survey item.
The survey item list must contain a `text` key (question text) and the following optional keys:
`label` for an item identifier; `type` for the answer type (one of
`"learnr_radio", "learnr_checkbox", "learnr_text", "learnr_numeric"`); `question_args` as list of
optional arguments passed to `learnr::question`; `answer_fn` as answer function (only considered
in case of `"learnr_numeric"` type); `answers` as list of answer options (only considered in case of
`"learnr_radio"` or `"learnr_checkbox"` type) – each answer option must in turn be a list with the key
`text` (answer option text) an the optional key `label` (answer option identifier).}

\item{caption}{Optional survey caption (defaults to "Survey")}

\item{message}{Optional message to display when survey item was submitted (defaults to "Thank you.")}
}
\value{
A survey as `tutorial_quiz` class list.
}
\description{
Generate a survey from survey item definitions passed as list `items`.
}
