% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{survey_likert}
\alias{survey_likert}
\title{Generate a survey with Likert scale type answers.}
\usage{
survey_likert(items, levels, caption = "Survey", message = "Thank you.")
}
\arguments{
\item{items}{A character vector with item questions. Can be a named vector – in this case the names are taken as
labels (identifiers) for the survey items.}

\item{levels}{Likert scale levels. Can be a character vector or list. If it is a character vector, we assume that
`levels` defines the same Likert scale levels for *all* survey items. Then this vector represents
the answer options. It can be a named vector, in which case the names represent the answer values.
If no names are given, values from 1 to `length(levels)` are generated. If this parameter is a list
it must have the same length as `items` and hence represents the answer options *per survey item*.
In this case each list entry must be a (optionally named) character vector.}

\item{caption}{Optional survey caption (defaults to "Survey")}

\item{message}{Optional message to display when survey item was submitted (defaults to "Thank you.")}
}
\value{
A survey as `tutorial_quiz` class list.
}
\description{
Generate a survey with Likert scale type answers.
}
