% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{use_learnrextra}
\alias{use_learnrextra}
\title{Use this package in a shiny app}
\usage{
use_learnrextra(consentmodal = TRUE, dataprotectmodal = TRUE)
}
\arguments{
\item{consentmodal}{either logical value or string pointing to an HTML file for a "data protection consent" modal}

\item{dataprotectmodal}{either logical value or string pointing to an HTML file for a "data protection license" modal}
}
\value{
tag list with sources to include
}
\description{
Enables the tracking functionality in a shiny app.
}
